Bug Fixes
•FIX: Mess counter correctly displays the number of known messes. (Note:
this is based on the mess's "exists" attribute)
•FIX: Time Remaining no longer always displays 50.
•CHANGE: "Messes Remaining" title is now labeled "Known Messes" (Note:
Be sure to set the Shitstorm and messTemplate to have an "exists" value
of 0 in final build) because messes are mostly generated as the player
performs actions.
•CHANGE: Removed unnecessary indentation and new lines at the start of
some description lines.
•ADD: Intro scenes now have nametags.
•ADD: Somes scenes with dialogue or thoughts lacked the respective
sign-posting (quotes or braces). Added where noticed.
•ADD: Set some timeUsed attributes for some actions. Picking up or
dropping items and continuing a scene takes no time. Actions that take
place over multiple scenes subtract their timeUsed at the start of the
first scene. Changing rooms still use the default timeUsed (0.2) because
they are unassigned.