Standardization and Reference Fixes
•FIX: Fixed some more odd spacing and indentation in some scene
descriptions.
•FIX: Examining an inventory item now works properly for all inventory
slots again. This was broken with the addition of item stacking. The
method that calls the item description now checks a condensed version of
the itemList instead of the full itemList.
•FIX: Added reference to Closet3 to Hallway11.
•FIX: Changed Hallway7's incorrect reference to Hallway6 to Hallway10.
•FIX: Changed Closet1's incorrect reference to Hallway10 to Hallway6.
•FIX: Moved the Glutton to ServantPassage3, since ServantPassage2 was
unreachable.
•FIX: Added reference to Closet4 to Hallway21.
•REMOVE: Removed the test(beta) option from Landing1F, Landing3F,
Bathroom2, Closet1, and Kitchen.
•ADD: Floor number is now displayed in the location title. Resized text
box to compensate.
•ADD: Added 2 more inventory buttons for a total of 10. Slightly resized
options, inventory, and description panes to compensate.
•CHANGE: Standardized directions. The Landing is at the north end of the
manor.
•CHANGE: Changed to order that movement options appear in the list of
options. They now appear in the order of north, east, south, west.
•CHANGE: More than doubled the amount of starting time.
•CHANGE: The Trophy Room scene with the gorilla maid now checks if you
have 2 loaves of bread in your inventory (loaves2) rather than having
the first 2 inventory slots filled with bread (invblk2). The old method
didn't take into account the previous change of item stacking.
•CHANGE: The number of loaves of bread that the gorilla maid requires
being fed to appear in the Trophy Room has been reduced to 7 (from 8).
The gorilla maid still requires 2 loaves in your inventory for this
scene to trigger. Previous changes created a limit of 10 loaves of
bread. The gorilla maid's scene wasn't updated to reflect this change,
so resolving both the Trophy Room and Orchestra Pit messes with bread
wasn't possible until now. The Orchestra Pit can't be resolved with
bread if the gorilla maid's feed-count has been satisfied first, since
the Orchestra Pit scene requires 7 loaves in your inventory to trigger.
•TODO: Find another way to implement that additionalDescriptions that
describe the Air Freshener and Bucket in Bathroom 1 and 2, respectively.
They cause a blank option to appear before the "Examine _____ "
option(s). The conditions that supply these additionalDescriptions are
currently commented out.
•TODO: Move Shears and Wet Book. They're both used to reduce the amount
of time that it takes to disarm the Tripwire mess, but neither can be
reached until the Tripwire has already been disabled. This makes it
impossible to use the Oil Slick mess against Bratticus.
•TODO: Tweak timeUsed values for encounters and messes.
•BUG: Cleaning up the Oil Spill in Landing2F before the Wet Floor in
Bathroom1 makes it impossible to clean the Wet Floor.